Tue, Oct 24, 2023, 5:27 AM +02:00

+++ recover/recover.c
@@ -43,7 +43,7 @@ int main(int argc, char *argv[])
 
             memo_block = 1;
 
-            int* outfile_ptr[counter_image] = (int*) malloc(memo_block * sizeof(BLOCK_SIZE));
+            int* outfile_ptr[counter_image] = (BYTE*) malloc(memo_block * sizeof(BLOCK_SIZE));
             if (outfile_ptr == NULL)
             {
                 printf("Memory not allocated for output file"\n);
